+++ CHANGELOG.md
@@ -7,6 +7,13 @@ The change log is available [on GitHub][2].
 [1]: http://semver.org/spec/v2.0.0.html
 [2]: https://github.com/roman/Haskell-teardown/libraries/teardown/CHANGELOG.md
 
+## v0.3.0.0 [PRE-RELEASE]
+
+* Bump from lts-9.1 to lts-9.5
+* Add `Control.Monad.Component` module
+* Add `IResource` instance for `[Teardown]`
+
+
 ## v0.2.0.0
 
 * Bump from lts-8.21 to lts-9.1